+++ src/main/java/com/yidusoft/project/transaction/dao/ProductSettingsMapper.java
@@ -3,11 +3,12 @@ package com.yidusoft.project.transaction.dao;
 import com.yidusoft.core.Mapper;
 import com.yidusoft.project.transaction.domain.ProductSettings;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductSettingsMapper extends Mapper<ProductSettings> {
 
     List<ProductSettings> getProductAll();
 
-    Integer getUnitPriceByTotal();
+    BigDecimal getUnitPriceByTotal();
 }
