+++ src/main/java/com/yidusoft/project/transaction/service/impl/ProductSettingsServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -26,7 +27,7 @@ public class ProductSettingsServiceImpl extends AbstractService<ProductSettings>
     }
 
     @Override
-    public Integer getUnitPriceByTotal() {
+    public BigDecimal getUnitPriceByTotal() {
         return productSettingsMapper.getUnitPriceByTotal();
     }
 }
